feat: add special configuration for exchange service
