+++ packages/manager/modules/hub/src/components/product-list/controller.js
@@ -1,3 +1,9 @@
 import { ListLayoutHelper } from '@ovh-ux/manager-ng-layout-helpers';
 
-export default class ManagerHubBillingProductList extends ListLayoutHelper.ListLayoutCtrl {}
+export default class ManagerHubBillingProductList extends ListLayoutHelper.ListLayoutCtrl {
+  $onInit() {
+    this.datagridId = `dg-${this.productType}`;
+
+    super.$onInit();
+  }
+}
